semver: move scylla version helpers out of the api
Those are not API types so they don't belong there and broke deep copy
generation. Also helpers shouldn't be coupled to the API defs.
